Completed first test
This is our first test - it tests to make sure that the
allFeeds variable has been defined and that it is not
empty. Experiment with this before you get started on
the rest of this project. What happens when you change
allFeeds in app.js to be an empty array and refresh the
page?

It fails the test.